backend/requirements.txt
sqlalchemy
flask
flask-restful
mysql-connector-python
psycopg[binary]
flask-sqlalchemy
python-dotenv
